+++ src/lib.rs
@@ -13,6 +13,7 @@ mod ref_count;
 mod runtime;
 mod runtime_name;
 mod runtime_type;
+mod try_into;
 mod unknown;
 mod weak;
 
@@ -29,6 +30,7 @@ pub use ref_count::*;
 pub use runtime::*;
 pub use runtime_name::*;
 pub use runtime_type::*;
+pub use try_into::*;
 pub use unknown::*;
 pub use winrt_macros::*;
