+++ src/unknown.rs
@@ -71,7 +71,7 @@ pub fn safe_query<From: ComInterface, Into: ComInterface>(from: &From) -> Into {
 }
 
 #[repr(C)]
-pub struct abi_IUnknown {
+pub(crate) struct abi_IUnknown {
     pub query: extern "system" fn(RawPtr, &Guid, *mut RawPtr) -> ErrorCode,
     addref: extern "system" fn(RawPtr) -> u32,
     release: extern "system" fn(RawPtr) -> u32,
